ada: Fix printing of numbers in JSON output for data representation
When calling GNAT with -gnatRj to generate JSON output for the
data representation of types and objects, it could happen that
numbers are printed in the Ada syntax for hexadecimal numbers, which
leads to an invalid JSON file being generated. Now fixed both for
the JSON output and the Ada-like output.

gcc/ada/

	* repinfo.adb (Compute_Max_Length): Set parameter to print number
	in decimal notation.
	(List_Component_Layout): Same.

Tested on x86_64-pc-linux-gnu, committed on master.